refactor: reuse seven-year date validation
