+++ web/src/utils/calendar/index.ts
@@ -6,6 +6,7 @@ export type {
 } from "./CalendarAdapter";
 export {
   DEFAULT_SEVEN_YEAR_ANCHOR_DATE,
+  isLocalDateString,
   parseLocalDateString,
 } from "./CalendarAdapter";
 export {
